app/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.4
huggingface-hub==0.26.2
langchain==0.3.4
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-openai==0.2.4
langchain-text-splitters==0.3.0
sentence-transformers==3.2.1
transformers==4.45.2
torch==2.5.0
openvino==2024.4.1.dev20240926
optimum==1.23.2
optimum-intel==1.20.0
tensorflow==2.18.0
optimum[openvino]
tf_keras==2.18.0
pypdf==5.1.0
chromadb==0.5.16
streamlit==1.39.0
unstructured[pdf]
langchain-unstructured[local]

